+++ hrx-delivery-woo/classes/Cronjob.php
@@ -7,7 +7,8 @@ if ( ! defined('ABSPATH') ) {
 }
 
 use HrxDeliveryWoo\Core;
-use HrxDeliveryWoo\Terminal;
+use HrxDeliveryWoo\Warehouse;
+use HrxDeliveryWoo\LocationsDelivery;
 use HrxDeliveryWoo\Debug;
 
 class Cronjob
@@ -121,12 +122,19 @@ class Cronjob
 
     public function job_update_delivery_locs()
     {
-        $status = Terminal::update_delivery_locations();
-
+        $status = LocationsDelivery::update_couriers();
         if ( $status['status'] == 'OK' ) {
-            $debug_msg = 'Successfully updated delivery locations. Added ' . $status['added'] . ', updated ' . $status['updated'] . ', failed ' . $status['failed'];
+            $debug_msg = 'Successfully updated courier delivery locations. Added ' . $status['added'] . ', updated ' . $status['updated'] . ', failed ' . $status['failed'];
         } else {
-            $debug_msg = 'Failed to update delivery locations. Error: ' . $status['msg'];
+            $debug_msg = 'Failed to update courier delivery locations. Error: ' . $status['msg'];
+        }
+        Debug::to_log($debug_msg, 'cronjob', true);
+
+        $status = LocationsDelivery::update(false);
+        if ( $status['status'] == 'OK' ) {
+            $debug_msg = 'Successfully updated terminal delivery locations. Added ' . $status['added'] . ', updated ' . $status['updated'] . ', failed ' . $status['failed'];
+        } else {
+            $debug_msg = 'Failed to update terminal delivery locations. Error: ' . $status['msg'];
         }
         Debug::to_log($debug_msg, 'cronjob', true);
     }