+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased]
+### Improved
+- Changed get of delivery countries from API instead of code
+- Added option to select shipping method title type on Checkout page
+
+### Updated
+- API library to v1.0.3
+
 ## [1.1.0.1] - 2023-07-07
 ### Fixed
 - Fixed delivery locations Update button action